requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
gunicorn==20.1.0
Pillow==9.4.0
requests==2.28.2
PyYAML==6.0
gevent==21.12.0
APScheduler==3.9.1
beautifulsoup4==4.12.3
lxml==4.9.4
psutil==5.9.5
